Job: UsersCleanerJob unban update

+++ LabsQueueBot.Core/Settings/TelegramBotSettings.cs
@@ -10,7 +10,7 @@ public class TelegramBotSettings
     public required IEnumerable<long> AdminChatId { get; init; } = [];
     public required TimeSpan UnionTimeUtc { get; init; }
     public required int LocalUtcOffset { get; init; }
-    public required double CleanerJobTimeoutInMinutes { get; init; }
+    public required int CleanerJobTimeoutInMinutes { get; init; }
     public required int StateUpdateTimeoutInMinutes { get; init; }
     public required int StateAllowedIntervalInMinutes { get; init; }
     public required int MaxBanTimeoutInDays { get; init; } = GlobalConstants.DefaultMaxBanTimeoutInDays;
